normalize script name: tools/dist
The name of scripts should not have a language specific file
extension, unless the script's purpose actually is illustrating how to
write something in that particular language.

This makes it possible to later change the implemantion language of
the script without needing to update callers, documentation, etc. with
a changed script file name extension.

Relates-to: https://github.com/lack/soundcraft-utils/issues/32