it: install files in both python and shell script
For installing and removing files and similar actions potentially
requiring special access priviledges, provide the objects
representing the files to be installed/removed with a second
implementation of the action:

  * keep the old way of running unpriviledged Python code, but for
    the case when that fails due to lack of priviledges,

  * add a way to add a shell command to a shell script to be audited
    by the user and then run with elevated priviledges, e.g. via sudo

Going with simple shell commands should make auditing the shell
script easier than e.g. creating a Python script.

Why are we not getting rid of the Python implementation and replacing
that with just invoking the shell commands? Good question.