it: Restructure into classes, directly install XDG files
Restructure the installation parts into classes: One class for
installing/uninstalling the D-Bus service, and one class for
installing/uninstalling the XDG desktop file and icons.

Introduce templating to the `*.desktop` file to install it with
the correct Gtk app id and a full path to the `soundcraft_gui`
executable.

Stop running `xdg-*` programs for installing XDG desktop and icon
files and directly install the files in the appropriate directory
instead.

Also, stop installing the 256px sized PNG icon file as if it were
of size 16, 24, 32, and 48.

Note: I am not sure whether directly installing the XDG files is
the correct way, though. It might be that `soundcraft_installtool`
is supposed to call some `xdg-*` program after all. It might also
be that whoever is calling `soundcraft_installtool` (a user or a
packager) is supposed to call some `xdg-*` program after
`soundcraft_installtool` has finished. We will have to find that
out and change things later, if necessary.