Add man pages for executables used at run time
Add man pages for the executables used at run time

  * socranop-ctl(1)

  * socranop-gui(1)

  * socranop-session-service(1)
    Yes, section 1 as it is user runnable.

and (un)install them with `socranop-installtool`.

There is no man page for `socranop-installtool`, as that
is not used after package installation and prepared
packages are expected not to ship it. Besides, that man
page is only available after once properly running
`socranop-installtool --post-install` without a man page.

Closes: https://github.com/socratools/socranop/issues/8
Closes: https://github.com/socratools/socranop/issues/22