Wait for the D-Bus service to leave the bus after shutdown
If we don't wait for the bus to be clear, there's a race against the
possibility that the next client we start up will see this nearly-gone
service instead of launching a new one as we intend.  The client will
also throw exceptions if that happens as the old service is removed from
the bus.