doc: make PERMISSIONS.md the source for socranop-permissions(7)
Make PERMISSIONS.md the source for the socranop-permissions(7)
man page and regenerate the socranop-permissions.7 file using
pandoc.

Corollaries:

  * pre-commit now runs pandoc when necessary
  * The pandoc utility needs to be installed when working on
    PERMISSIONS.md

However, we have NOT added the pre-commit check to the github
workflow commit-checks on purpose. Different pandoc versions
create different man page files, therefore github's workflow
often generates a man page file which differs from the one
generated on the developer's workstation which is part of the
commit. And that will break the github workflow commit-checks.

We do not have such a rule to generate PNG files from the SVG icon,
so this is not the first such script not updating its generated
files on github workflow commit-checks.