Correct spelling mistakes in code comments
Correct a number of spelling mistakes in code comments.

This also adds a pre-commit hook to keep the dictionary items
inside the custom private dictionary file sorted and unique.

This does NOT automatically add a pre-commit hook checking the
spelling of all comments, as that check requires additional
software to be installed.