Pipfile: Completely remove pipenv-setup
The pipenv-setup main branch has not seen a commit since 2022-06-06,
the latest release 3.2.0 is from 2022-02-22, and pipenv-setup is
incompatible with some of the packages it pulls in (e.g. plette).

The https://pypi.org/project/pipenv-setup/ page badge says "Python
2.7|3.5|3.6|3.7|3.8" while socranop supports 3.9, 3.10, 3.11, 3.12.

We might want to re-add the pipenv-setup functionality later
(keeping Pipfile and setup.py dependency versions in sync), when/if
pipenv-setup becomes maintained again. Until then, pipenv-setup
causes more trouble than it solves.

It was the fault of pipenv-setup that we needed limit vistir to
an older version. As we are removing pipenv-setup, we can also
remove that old version requirements as well. That includes the
whole requirement on vistir, which we only needed to limit its
version.

We must keep tox at 3.x (before 4.0) for other reasons.

This means we now need to explicitly require wheel.

PF