gui: Add "About socranop" button to the HeaderBar
Some systems were showing the app menu and some were not.

This means we need to replace the app menu with its "About" and "Quit"
menu items with something portable: A Gtk.HeaderBar.

This makes the "About socranop" dialog reachable from all systems.

Note that we now explicitly have the audio routing `Gtk.ComboBox`
grab the keyboard focus, as the implicit focus on the first focusable
widget would now focus the "About socranop" button in the HeaderBar.

This also updates the screenshot images used for README.md and
(later) for appstream info and other things.

Fixes: https://github.com/socratools/socranop/issues/87