gui: Add keyboard accels for About and Quit
Add keyboard accelerators for the "About socranop" dialog
and for quitting the application: Ctrl+B (**b**bout) and
Ctrl+Q, respectively.

Do not use the letter "a" as that is currently already
associated with "Apply" (as Alt+A).

Using both Alt+* and Ctrl+* makes the keyboard shortcuts a
bit inconsistent, but at least there are keyboard shortcuts
for everything.

Fixes: https://github.com/socratools/socranop/issues/73